split db schema into multiple files
